Use assistant builder context instead of props drilling
